[Task 9] - Service Analyzer/requirements.txt
# Core dependencies
aiohttp>=3.8.0,<4.0.0
typer>=0.9.0,<1.0.0
rich>=13.0.0,<14.0.0
python-dotenv>=1.0.0,<2.0.0
pydantic>=2.0.0,<3.0.0

# Testing dependencies
pytest>=7.0.0,<8.0.0
pytest-asyncio>=0.21.0,<1.0.0
pytest-cov>=4.0.0,<5.0.0
pytest-mock>=3.10.0,<4.0.0

# Development dependencies
black>=23.0.0,<24.0.0
isort>=5.12.0,<6.0.0
flake8>=6.0.0,<7.0.0
mypy>=1.0.0,<2.0.0
pre-commit>=3.0.0,<4.0.0 